fix: fix errors

+++ Minibank.Core/Domains/Users/Services/UserService.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.DependencyInjection;
+using Minibank.Core.Domains.BankAccounts.Repositories;
 using Minibank.Core.Domains.Users.Repositories;
 
 namespace Minibank.Core.Domains.Users.Services
@@ -9,10 +10,12 @@ namespace Minibank.Core.Domains.Users.Services
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepository;
-        
-        public UserService(IUserRepository userRepository)
+
+        private readonly IBankAccountRepository _bankAccountRepository;
+        public UserService(IUserRepository userRepository, IBankAccountRepository bankAccountRepository)
         {
             _userRepository = userRepository;
+            _bankAccountRepository = bankAccountRepository;
         }
         public User GetUser(string id)
         {
@@ -25,19 +28,12 @@ namespace Minibank.Core.Domains.Users.Services
             if (user.Login.Contains(" ")) throw new ValidationException("Login must not have some spaces!");
             if (user.Login.Length > 20)
                 throw new ValidationException("Login's length must not be more that 20 symbols!");
-            _userRepository.CreateUser(user);
+            _userRepository.CreateUser(user.Login, user.Email);
         }
 
         public IEnumerable<User> GetAllUsers()
         {
-            return _userRepository.GetAllUsers()
-                .Select(it => new User
-                {
-                    Id = it.Id,
-                    Login = it.Login,
-                    Email = it.Email,
-                    HasBankAccounts = it.HasBankAccounts
-                });
+            return _userRepository.GetAllUsers();
         }
 
         public void UpdateUser(User user)
@@ -47,18 +43,10 @@ namespace Minibank.Core.Domains.Users.Services
 
         public void DeleteUser(string id)
         {
+            if (_bankAccountRepository.HasBankAccounts(id))
+                throw new ValidationException("You can't delete user which have one or more BanAccounts");
             _userRepository.DeleteUser(id);
         }
-        
-        
 
-        // public void SetActive(string id)
-        // {
-        //     var user = _userRepository.Get(id);
-        //     if (user == null) return;
-        //     if (user.IsActive) throw new ValidationException("User has already active!");
-        //     user.IsActive = true;
-        //     _userRepository.Update(user);
-        // }
     }
 }
